Remove ssize_t (#10664)

+++ native/cocos/base/ZipUtils.cpp
@@ -63,7 +63,7 @@ bool         ZipUtils::encryptionKeyIsValid = false;
 
 // --------------------- ZipUtils ---------------------
 
-inline void ZipUtils::decodeEncodedPvr(unsigned int *data, ssize_t len) {
+inline void ZipUtils::decodeEncodedPvr(unsigned int *data, uint32_t len) {
     const int enclen    = 1024;
     const int securelen = 512;
     const int distance  = 64;
@@ -126,7 +126,7 @@ inline void ZipUtils::decodeEncodedPvr(unsigned int *data, ssize_t len) {
     }
 }
 
-inline unsigned int ZipUtils::checksumPvr(const unsigned int *data, ssize_t len) {
+inline unsigned int ZipUtils::checksumPvr(const unsigned int *data, uint32_t len) {
     unsigned int cs    = 0;
     const int    cslen = 128;
 
@@ -143,11 +143,11 @@ inline unsigned int ZipUtils::checksumPvr(const unsigned int *data, ssize_t len)
 // Should buffer factor be 1.5 instead of 2 ?
 #define BUFFER_INC_FACTOR (2)
 
-int ZipUtils::inflateMemoryWithHint(unsigned char *in, ssize_t inLength, unsigned char **out, ssize_t *outLength, ssize_t outLengthHint) {
+int ZipUtils::inflateMemoryWithHint(unsigned char *in, uint32_t inLength, unsigned char **out, uint32_t *outLength, uint32_t outLengthHint) {
     /* ret value */
-    int     err        = Z_OK;
-    ssize_t bufferSize = outLengthHint;
-    *out               = static_cast<unsigned char *>(malloc(bufferSize));
+    int     err         = Z_OK;
+    uint32_t bufferSize = outLengthHint;
+    *out                = static_cast<unsigned char *>(malloc(bufferSize));
 
     z_stream descompressionStream; /* decompression stream */
     descompressionStream.zalloc = static_cast<alloc_func>(nullptr);
@@ -155,9 +155,9 @@ int ZipUtils::inflateMemoryWithHint(unsigned char *in, ssize_t inLength, unsigne
     descompressionStream.opaque = static_cast<voidpf>(nullptr);
 
     descompressionStream.next_in   = in;
-    descompressionStream.avail_in  = static_cast<unsigned int>(inLength);
+    descompressionStream.avail_in  = inLength;
     descompressionStream.next_out  = *out;
-    descompressionStream.avail_out = static_cast<unsigned int>(bufferSize);
+    descompressionStream.avail_out = bufferSize;
 
     /* window size to hold 256k */
     if ((err = inflateInit2(&descompressionStream, 15 + 32)) != Z_OK) {
@@ -202,9 +202,9 @@ int ZipUtils::inflateMemoryWithHint(unsigned char *in, ssize_t inLength, unsigne
     return err;
 }
 
-ssize_t ZipUtils::inflateMemoryWithHint(unsigned char *in, ssize_t inLength, unsigned char **out, ssize_t outLengthHint) {
-    ssize_t outLength = 0;
-    int     err       = inflateMemoryWithHint(in, inLength, out, &outLength, outLengthHint);
+uint32_t ZipUtils::inflateMemoryWithHint(unsigned char *in, uint32_t inLength, unsigned char **out, uint32_t outLengthHint) {
+    uint32_t outLength = 0;
+    int     err        = inflateMemoryWithHint(in, inLength, out, &outLength, outLengthHint);
 
     if (err != Z_OK || *out == nullptr) {
         if (err == Z_MEM_ERROR) {
@@ -227,7 +227,7 @@ ssize_t ZipUtils::inflateMemoryWithHint(unsigned char *in, ssize_t inLength, uns
     return outLength;
 }
 
-ssize_t ZipUtils::inflateMemory(unsigned char *in, ssize_t inLength, unsigned char **out) {
+uint32_t ZipUtils::inflateMemory(unsigned char *in, uint32_t inLength, unsigned char **out) {
     // 256k for hint
     return inflateMemoryWithHint(in, inLength, out, 256 * 1024);
 }
@@ -307,8 +307,8 @@ bool ZipUtils::isCCZFile(const char *path) {
     return isCCZBuffer(compressedData.getBytes(), compressedData.getSize());
 }
 
-bool ZipUtils::isCCZBuffer(const unsigned char *buffer, ssize_t len) {
-    if (static_cast<size_t>(len) < sizeof(struct CCZHeader)) {
+bool ZipUtils::isCCZBuffer(const unsigned char *buffer, uint32_t len) {
+    if (len < sizeof(struct CCZHeader)) {
         return false;
     }
 
@@ -328,7 +328,7 @@ bool ZipUtils::isGZipFile(const char *path) {
     return isGZipBuffer(compressedData.getBytes(), compressedData.getSize());
 }
 
-bool ZipUtils::isGZipBuffer(const unsigned char *buffer, ssize_t len) {
+bool ZipUtils::isGZipBuffer(const unsigned char *buffer, uint32_t len) {
     if (len < 2) {
         return false;
     }
@@ -336,7 +336,7 @@ bool ZipUtils::isGZipBuffer(const unsigned char *buffer, ssize_t len) {
     return buffer[0] == 0x1F && buffer[1] == 0x8B;
 }
 
-int ZipUtils::inflateCCZBuffer(const unsigned char *buffer, ssize_t bufferLen, unsigned char **out) {
+int ZipUtils::inflateCCZBuffer(const unsigned char *buffer, uint32_t bufferLen, unsigned char **out) {
     const auto *header = reinterpret_cast<const struct CCZHeader *>(buffer);
 
     // verify header
@@ -372,8 +372,8 @@ int ZipUtils::inflateCCZBuffer(const unsigned char *buffer, ssize_t bufferLen, u
 
 #if CC_DEBUG > 0
         // decrypt
-        auto *  ints   = reinterpret_cast<unsigned int *>(const_cast<unsigned char *>(buffer) + 12);
-        ssize_t enclen = (bufferLen - 12) / 4;
+        auto *  ints    = reinterpret_cast<unsigned int *>(const_cast<unsigned char *>(buffer) + 12);
+        uint32_t enclen = (bufferLen - 12) / 4;
 
         decodeEncodedPvr(ints, enclen);
         // verify checksum in debug mode
@@ -548,7 +548,7 @@ bool ZipFile::fileExists(const ccstd::string &fileName) const {
     return ret;
 }
 
-unsigned char *ZipFile::getFileData(const ccstd::string &fileName, ssize_t *size) {
+unsigned char *ZipFile::getFileData(const ccstd::string &fileName, uint32_t *size) {
     unsigned char *buffer = nullptr;
     if (size) {
         *size = 0;
@@ -576,7 +576,7 @@ unsigned char *ZipFile::getFileData(const ccstd::string &fileName, ssize_t *size
         CCASSERT(nSize == 0 || nSize == (int)fileInfo.uncompressed_size, "the file size is wrong");
 
         if (size) {
-            *size = fileInfo.uncompressed_size;
+            *size = static_cast<uint32_t>(fileInfo.uncompressed_size);
         }
         unzCloseCurrentFile(*zipFile);
     } while (false);
