+++ native/cocos/platform/Image.h
@@ -61,10 +61,10 @@ public:
     };
 
     bool initWithImageFile(const ccstd::string &path);
-    bool initWithImageData(const unsigned char *data, ssize_t dataLen);
+    bool initWithImageData(const unsigned char *data, uint32_t dataLen);
 
     // @warning kFmtRawData only support RGBA8888
-    bool initWithRawData(const unsigned char *data, ssize_t dataLen, int width, int height, int bitsPerComponent, bool preMulti = false);
+    bool initWithRawData(const unsigned char *data, uint32_t dataLen, int width, int height, int bitsPerComponent, bool preMulti = false);
 
     // data will be free outside.
     inline void takeData(unsigned char **outData) {
@@ -74,7 +74,7 @@ public:
 
     // Getters
     inline unsigned char *getData() const { return _data; }
-    inline ssize_t        getDataLen() const { return _dataLen; }
+    inline uint32_t       getDataLen() const { return _dataLen; }
     inline Format         getFileType() const { return _fileType; }
     inline gfx::Format    getRenderFormat() const { return _renderFormat; }
     inline int            getWidth() const { return _width; }
@@ -84,18 +84,18 @@ public:
     inline bool isCompressed() const { return _isCompressed; }
 
 protected:
-    bool initWithJpgData(const unsigned char *data, ssize_t dataLen);
-    bool initWithPngData(const unsigned char *data, ssize_t dataLen);
-    bool initWithWebpData(const unsigned char *data, ssize_t dataLen);
-    bool initWithPVRData(const unsigned char *data, ssize_t dataLen);
-    bool initWithPVRv2Data(const unsigned char *data, ssize_t dataLen);
-    bool initWithPVRv3Data(const unsigned char *data, ssize_t dataLen);
-    bool initWithETCData(const unsigned char *data, ssize_t dataLen);
-    bool initWithETC2Data(const unsigned char *data, ssize_t dataLen);
-    bool initWithASTCData(const unsigned char *data, ssize_t dataLen);
+    bool initWithJpgData(const unsigned char *data, uint32_t dataLen);
+    bool initWithPngData(const unsigned char *data, uint32_t dataLen);
+    bool initWithWebpData(const unsigned char *data, uint32_t dataLen);
+    bool initWithPVRData(const unsigned char *data, uint32_t dataLen);
+    bool initWithPVRv2Data(const unsigned char *data, uint32_t dataLen);
+    bool initWithPVRv3Data(const unsigned char *data, uint32_t dataLen);
+    bool initWithETCData(const unsigned char *data, uint32_t dataLen);
+    bool initWithETC2Data(const unsigned char *data, uint32_t dataLen);
+    bool initWithASTCData(const unsigned char *data, uint32_t dataLen);
 
     unsigned char *_data     = nullptr;
-    ssize_t        _dataLen  = 0;
+    uint32_t       _dataLen  = 0;
     int            _width    = 0;
     int            _height   = 0;
     Format         _fileType = Format::UNKNOWN;
@@ -103,14 +103,14 @@ protected:
     ccstd::string  _filePath;
     bool           _isCompressed = false;
 
-    static Format detectFormat(const unsigned char *data, ssize_t dataLen);
-    static bool   isPng(const unsigned char *data, ssize_t dataLen);
-    static bool   isJpg(const unsigned char *data, ssize_t dataLen);
-    static bool   isWebp(const unsigned char *data, ssize_t dataLen);
-    static bool   isPvr(const unsigned char *data, ssize_t dataLen);
-    static bool   isEtc(const unsigned char *data, ssize_t dataLen);
-    static bool   isEtc2(const unsigned char *data, ssize_t dataLen);
-    static bool   isASTC(const unsigned char *data, ssize_t detaLen);
+    static Format detectFormat(const unsigned char *data, uint32_t dataLen);
+    static bool   isPng(const unsigned char *data, uint32_t dataLen);
+    static bool   isJpg(const unsigned char *data, uint32_t dataLen);
+    static bool   isWebp(const unsigned char *data, uint32_t dataLen);
+    static bool   isPvr(const unsigned char *data, uint32_t dataLen);
+    static bool   isEtc(const unsigned char *data, uint32_t dataLen);
+    static bool   isEtc2(const unsigned char *data, uint32_t dataLen);
+    static bool   isASTC(const unsigned char *data, uint32_t detaLen);
 
     static gfx::Format getASTCFormat(const unsigned char *pHeader);
